WIP: Add `forall a. Show (t a)` superclass to `GShow`
As `GShow` is isomorphic to this constraint, it's nice to add this super
class. This reminds people that if they implement `GShow` they should
also implement `forall a. Show (t a)`, as its both incredibly easy to do
so and never (in practice) an additional dependency.

Unforunately, this change doesn't yet type-check because the instances
in base for `Sum` and `Product` are overly consersative: they assume
there is a field with a type related to the GADT index. That should be
fixed, and this PR can be pointed to as an example of why the problem
isn't just theoretical.